backend/pyproject.toml
[project]
name = "retrievai-backend"
version = "0.2.0"
description = "RetrievAI Backend API - Async FastAPI with ARQ"
requires-python = ">=3.12"
dependencies = [
    # FastAPI ecosystem
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.12",

    # Database (async)
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",

    # Security
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.1",
    "email-validator>=2.1.0",

    # Background tasks
    "arq>=0.26.0",
    "redis>=5.2.0",

    # LangChain & RAG
    "langchain>=0.3.7",
    "langchain-openai>=0.2.8",
    "langchain-community>=0.3.7",
    "langchain-chroma>=0.1.4",
    "langchain-text-splitters>=0.3.0",
    "chromadb>=0.5.18",
    "openai>=1.54.4",

    # Document processing
    "pymupdf>=1.24.13",
    "pymupdf4llm>=0.0.17",
    "ocrmypdf>=16.6.2",
    "wordsegment>=1.3.1",
    "nltk>=3.9.1",
    "unstructured>=0.16.0",

    # Utils
    "aiofiles>=24.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
